+++ src/main/kotlin/calculator/domain/CustomReadStrategy.kt
@@ -1,16 +1,18 @@
-package calculator
+package calculator.domain
+
+import calculator.interfaces.StringReadStrategy
 
 class CustomReadStrategy() : StringReadStrategy {
     override fun readString(string: String): List<String> {
         val separator = getCustomSeparator(string)
         val stringWithOutCustomSeparator = getStringWithOutCustomSeparator(string)
 
-        return stringWithOutCustomSeparator.split(separator)
+        return stringWithOutCustomSeparator.split(separator!!)
     }
 
-    private fun getCustomSeparator(string: String): String {
+    private fun getCustomSeparator(string: String): String? {
         val regexResult = Regex("//(.)\n(.*)").find(string)
-        return regexResult!!.groupValues[1]
+        return regexResult?.let {it.groupValues[1]}
     }
 
     private fun getStringWithOutCustomSeparator(string: String): String {
