+++ src/main/kotlin/calculator/domain/StandardReadStrategy.kt
@@ -1,4 +1,6 @@
-package calculator
+package calculator.domain
+
+import calculator.interfaces.StringReadStrategy
 
 class StandardReadStrategy() : StringReadStrategy {
     override fun readString(string: String): List<String> {
