+++ src/main/kotlin/calculator/interfaces/StringReadStrategy.kt
@@ -1,4 +1,4 @@
-package calculator
+package calculator.interfaces
 
 interface StringReadStrategy {
     fun readString(string: String): List<String>
